refactor: split book theme skin and presentation
