+++ common/src/main/java/com/klikli_dev/modonomicon/client/gui/book/theme/BookPresentationTheme.java
@@ -1,0 +1,14 @@
+/*
+ * SPDX-FileCopyrightText: 2026 klikli-dev
+ *
+ * SPDX-License-Identifier: MIT
+ */
+
+package com.klikli_dev.modonomicon.client.gui.book.theme;
+
+public interface BookPresentationTheme {
+
+    BookLayoutTheme layout();
+
+    BookPaletteTheme palette();
+}
