+++ common/src/main/java/com/klikli_dev/modonomicon/client/gui/book/theme/BookSkinTheme.java
@@ -1,0 +1,20 @@
+/*
+ * SPDX-FileCopyrightText: 2026 klikli-dev
+ *
+ * SPDX-License-Identifier: MIT
+ */
+
+package com.klikli_dev.modonomicon.client.gui.book.theme;
+
+public interface BookSkinTheme {
+
+    BookContentTheme content();
+
+    BookOverviewTheme overview();
+
+    BookNodeTheme node();
+
+    BookFrameTheme frame();
+
+    BookRecipeTheme recipes();
+}
